+++ pkg/bitesize/service.go
@@ -2,10 +2,9 @@ package bitesize
 
 import (
 	"fmt"
+	validator "gopkg.in/validator.v2"
 	"strconv"
 	"strings"
-
-	validator "gopkg.in/validator.v2"
 )
 
 // Service represents a single service and it's configuration,
@@ -23,6 +22,7 @@ type Service struct {
 	Requests     ContainerRequests       `yaml:"requests" validate:"requests"`
 	HealthCheck  *HealthCheck            `yaml:"health_check,omitempty"`
 	EnvVars      []EnvVar                `yaml:"env,omitempty"`
+	DeployedPods []Pod                   `yaml:"-"` //Ignore field when parsing bitesize yaml
 	Annotations  []Annotation            `yaml:"annotations,omitempty"`
 	Volumes      []Volume                `yaml:"volumes,omitempty"`
 	Options      map[string]string       `yaml:"options,omitempty"`