+++ pkg/util/k8s/k8s.go
@@ -94,6 +94,11 @@ func (c *Client) PVC() *PersistentVolumeClaim {
 	return &PersistentVolumeClaim{Interface: c.Interface, Namespace: c.Namespace}
 }
 
+// Pod builds Pod client
+func (c *Client) Pod() *Pod {
+	return &Pod{Interface: c.Interface, Namespace: c.Namespace}
+}
+
 // Ingress builds Ingress client
 func (c *Client) Ingress() *Ingress {
 	return &Ingress{Interface: c.Interface, Namespace: c.Namespace}
@@ -116,5 +121,12 @@ func (c *Client) ThirdPartyResource(kind string) *ThirdPartyResource {
 func listOptions() v1.ListOptions {
 	return v1.ListOptions{
 		LabelSelector: "creator=pipeline",
+	}
+}
+func logOptions() *v1.PodLogOptions {
+	return &v1.PodLogOptions{
+		//SinceSeconds: &[]int64{300}[0], //Gets last 5 minutes of logs
+		TailLines:  &[]int64{500}[0], //Retrieve last 500 lines from pod log
+		Timestamps: true,             //Add timestamp to each line in the log
 	}
 }